further modularization
